pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sakura-ml"
version = "0.1.2"
description = "Sakura provides asynchronous training for DNN."
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.10"
authors = [{ name = "ZakuroAI", email = "git@zakuro.ai" }]
dependencies = [
    "tqdm>=4.64.0",
    "torch",
    "torchvision",
    "numpy",
    "gnutools-python",
    "six",
    "chardet",
    "lightning",
    "ipython",
    "charset-normalizer==3.1.0",
    # Zakuro replaces the former MPI + Redis transport: per-epoch validation
    # is dispatched to a Zakuro worker via @zk.fn instead of a second MPI rank.
    # Pulled from git until the 0.2.x series is published to PyPI.
    "zakuro-ai @ git+https://github.com/zakuro-ai/zakuro.git@master",
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
]

[dependency-groups]
dev = ["pytest"]

[project.scripts]
sakura = "sakura.__main__:main"
sakura-benchmark = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["sakura"]

[tool.hatch.build.targets.wheel.force-include]
"sakura/config.yaml" = "sakura/config.yaml"
"main.py" = "main.py"
